LayoutTests/imported/w3c/web-platform-tests/html/syntax/parsing/template/additions-to-foster-parenting/template-is-a-foster-parent-element-expected.txt


PASS Template is a foster parent element. Test <table> immediately below <template>
PASS Template is a foster parent element. Test <template> element without <table>
